Rename `declare_type_safe` -> `declare_macro`
