+++ crates/forge/tests/data/declare_macro/Scarb.toml
@@ -1,5 +1,5 @@
 [package]
-name = "declare_type_safe"
+name = "declare_macro"
 version = "0.1.0"
 edition = "2024_07"
