+++ crates/forge/tests/data/declare_macro/src/hello_starknet.cairo
