+++ crates/forge/tests/data/declare_macro/src/lib.cairo
